audio-render/requirements.txt
audioread==3.0.1
blinker==1.6.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
ffmpeg-python==0.2.0
Flask==2.3.2
fonttools==4.53.0
future==1.0.0
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-texttospeech==2.16.3
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.4.2
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.2
matplotlib==3.9.0
more-itertools==8.10.0
msgpack==1.0.8
netifaces==0.11.0
numba==0.60.0
numpy==2.0.0
oauthlib==3.2.0
packaging==24.1
pika==1.2.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.26.1
pycparser==2.22
pydub==0.25.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
scikit-learn==1.5.0
scipy==1.14.0
SecretStorage==3.3.1
setuptools==70.1.1
six==1.16.0
soundfile==0.12.1
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
wadllib==1.3.6
Werkzeug==2.3.4
zipp==1.0.0
